[RUMF-828] resolve configuration cyclic dependencies
Most issues were related to types and constants
* create a new 'userConfiguration.types' file, to regroup user facing
configuration types
* move transport-related types in 'transport' file.
